split otu into parts
